requirements.txt
fastapi
uvicorn[standard]
gunicorn
pandas
numpy
dotenv
PyJWT